auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
pytest
email-validator
httpx